pyproject.toml
[tool.poetry]
name = "oatflake"
version = "0.1.0"
description = "OATFLAKE - A RAG-based assistant platform"
authors = ["Mars"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/blob/OATFLAKE"
packages = [{include = "scripts"}, {include = "api"}]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
fastapi = "^0.115.0"
uvicorn = "^0.34.0"
supabase = "^2.13.0"
langchain = "^0.3.19"
langchain-community = "^0.3.18"
faiss-cpu = "^1.10.0"
sentence-transformers = {version = "^3.4.1", optional = true}
pydantic = "^2.10.0"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.7.1"
tiktoken = "^0.9.0"
pandas = "^2.2.0"
unstructured = "^0.16.21"
python-multipart = "^0.0.20"
slack-bolt = {version = "^1.22.0", optional = true}
py2app = {version = "^0.28.6", optional = true}
PyQt6 = {version = "^6.6.1", optional = true}
pyngrok = "^7.2.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
ruff = "^0.3.0"
mypy = "^1.8.0"
black = "^24.0.0"

[tool.poetry.extras]
slack = ["slack-bolt"]
desktop = ["py2app", "PyQt6"]
embeddings = ["sentence-transformers"]

[tool.poetry.scripts]
app = "app:main"
process = "run_complete_processing:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I"]
ignore = []

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false